Extract formatting functions from stuck_detection.py into stuck_formatting.py
Moves 6 formatting functions (~253 lines) to a dedicated module for improved
modularity and testability. The instance method _format_intervention_summary
is made a public standalone function. All existing tests pass unchanged and
new tests cover the extracted format_intervention_summary function.

Closes #1886
